Refactor NFT's folder repo
